+++ ui/components.d.ts
@@ -12,6 +12,7 @@ declare module 'vue' {
     CreateLink: typeof import('./src/components/Drawer/components/SessionShare/widget/CreateLink.vue')['default']
     Drawer: typeof import('./src/components/Drawer/index.vue')['default']
     FileManagement: typeof import('./src/components/Drawer/components/FileManagement/index.vue')['default']
+    General: typeof import('./src/components/Drawer/components/General/index.vue')['default']
     Keyboard: typeof import('./src/components/Drawer/components/Keyboard/index.vue')['default']
     MainContent: typeof import('./src/components/Kubernetes/MainContent/index.vue')['default']
     NButton: typeof import('naive-ui')['NButton']
